requirements.txt
asgiref==3.7.2
black==23.7.0
click==8.1.7
colorama==0.4.6
Django==4.2.4
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
